requirements.txt
Django==4.2.23
django-distill==3.2.7
requests
Pillow